requirements.txt
# Core Django
Django==5.2.8

# REST API
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# Environment Management
python-decouple==3.8

# Image Processing
Pillow==10.4.0

# CORS Headers
django-cors-headers==4.4.0

# Database (PostgreSQL - install when needed for production)
# psycopg2-binary==2.9.9

# Development Tools
django-debug-toolbar==4.4.6

# Testing
pytest==8.3.2
pytest-django==4.8.0
